requirements.txt
# ============================================================================
# Markdown 处理工具集 V1.0 - Python 依赖包
# ============================================================================
# 
# 本文件列出了运行本项目所需的 Python 依赖包
# 使用方法: pip install -r requirements.txt
#
# ============================================================================

# ----------------------------------------------------------------------------
# 核心依赖 (必需)
# ----------------------------------------------------------------------------
# 用途: 显示命令行进度条，提供友好的处理进度反馈
# 所有模块都会用到此库
tqdm>=4.66.0

# ----------------------------------------------------------------------------
# GUI 界面支持 (可选)
# ----------------------------------------------------------------------------
# 用途: 为 feishu2md.py 提供图形界面支持
# 如果只使用命令行版本 (main.py)，可以不安装此依赖
# 注意: 此包仅用于检测依赖，实际不直接使用 python-docx 的功能
python-docx>=1.1.0

# ============================================================================
# 外部程序依赖 (必需安装)
# ============================================================================
#
# Pandoc - 文档转换核心引擎
# ----------------------------------------------------------------------------
# Pandoc 是一个独立的命令行工具，无法通过 pip 安装
# 必须单独安装后才能使用 docx_to_markdown.py 模块
#
# 安装方法：
#
#   Windows (推荐使用 Chocolatey):
#     choco install pandoc
#
#   Windows (备选方案 - Winget):
#     winget install --id=JohnMacFarlane.Pandoc -e
#
#   macOS (使用 Homebrew):
#     brew install pandoc
#
#   Linux (Ubuntu/Debian):
#     sudo apt-get update && sudo apt-get install -y pandoc
#
#   或从官网下载: https://pandoc.org/installing.html
#
# 验证安装:
#   pandoc --version
#
# ============================================================================

# ============================================================================
# 安装指南
# ============================================================================
#
# 方法1: 使用自动安装脚本 (强烈推荐)
# ----------------------------------------------------------------------------
#   python install_dependencies.py --auto
#
#   此脚本会自动:
#   - 检测 Python 包是否已安装
#   - 检测 Pandoc 是否可用
#   - 自动安装缺失的依赖
#
# 方法2: 手动安装
# ----------------------------------------------------------------------------
#   # 安装 Python 包
#   pip install -r requirements.txt
#
#   # 安装 Pandoc (根据系统选择)
#   choco install pandoc  # Windows
#   brew install pandoc   # macOS
#
# 方法3: 最小安装 (仅命令行)
# ----------------------------------------------------------------------------
#   pip install tqdm
#   choco install pandoc  # Windows
#
# ============================================================================

# ============================================================================
# 版本说明
# ============================================================================
# - tqdm: 4.66.0+ (需要较新版本以获得更好的进度条显示)
# - python-docx: 1.1.0+ (用于依赖检测)
# - Pandoc: 建议 2.0+ (核心转换引擎)
#
# ============================================================================
# 兼容性
# ============================================================================
# - Python: 3.7+
# - 操作系统: Windows 10/11, macOS 10.14+, Linux (Ubuntu 18.04+)
#
# ============================================================================
